updated the client structure to follow SRP and OCP
