+++ source_code/client/user_client/article_service.py
@@ -1,10 +1,16 @@
+from news_api import NewsAPIClient
+
+
 class ArticleService:
-    def __init__(self, api_client, get_user):
+    def __init__(self, api_client: NewsAPIClient, get_user):
         self.api_client = api_client
         self.get_user = get_user
 
     def get_saved_articles(self):
         return self.api_client.make_request('GET', 'user/articles/saved', current_user=self.get_user())
+    
+    def get_saved_articles(self):
+        return self.api_client.make_request('GET', 'user/article', current_user=self.get_user())
 
     def save_article(self, article_id):
         return self.api_client.make_request('POST', 'user/articles/save', {'article_id': article_id}, current_user=self.get_user())
