+++ source_code/client/user_client/handle_menu/saved_articles_menu.py
@@ -1,0 +1,32 @@
+from utils import print_menu, display_articles, get_valid_article_id
+from user_client.article_service import ArticleService
+
+
+class SavedArticlesMenu:
+    def __init__(self, user_menu):
+        self.article_service: ArticleService = user_menu.article_service
+        self.user_service = user_menu.user_service
+
+    def run(self):
+        print("\n--- Saved Articles ---")
+        response = self.article_service.get_saved_articles()
+        if not isinstance(response, list) or not response[0].get('success'):
+            print("Failed to load saved articles.")
+            return True
+
+        display_articles(response[0].get('articles', []))
+
+        while True:
+            choice = print_menu("Manage Saved Articles", ["Back", "Logout", "Delete Article"])
+            if choice == '1':
+                return True
+            elif choice == '2':
+                self.user_service.logout()
+                return False
+            elif choice == '3':
+                article_id = get_valid_article_id()
+                if article_id:
+                    result = self.article_service.delete_article(article_id)
+                    print(result.get('message', "Failed to delete article."))
+            else:
+                print("Invalid option.")
