+++ source_code/client/user_client/user_menu.py
@@ -1,7 +1,11 @@
-from datetime import datetime, date
-from utils import display_articles, get_date
-from utils import print_menu, get_valid_article_id
-from user_client.handle_menu.headlines import Headlines
+from user_client.handle_menu.headlines_menu import HeadlinesMenu
+from user_client.handle_menu.saved_articles_menu import SavedArticlesMenu
+from user_client.handle_menu.search_menu import SearchMenu
+from user_client.handle_menu.notifications_menu import NotificationsMenu
+from utils import print_menu
+from datetime import datetime
+
+
 class UserMenu:
     def __init__(self, api_client, get_current_user, set_current_user,
                  article_service, notification_service, user_service):
@@ -14,371 +18,30 @@ class UserMenu:
 
     def run_menu(self):
         response = True
-        while True and response:
-            self.__welcome_user()
+        while response:
+            self._welcome_user()
             choice = print_menu("Main Menu", [
                 "Headlines", "Saved Articles", "Search", "Notifications", "Logout"
             ])
 
             if choice == '1':
-                response = self.__headlines_menu()
+                response = HeadlinesMenu(self).run()
             elif choice == '2':
-                self.__saved_articles_menu()
+                response = SavedArticlesMenu(self).run()
             elif choice == '3':
-                self.__search_menu()
+                response = SearchMenu(self).run()
             elif choice == '4':
-                response = self.__notifications_menu()
+                response = NotificationsMenu(self).run()
             elif choice == '5':
                 self.user_service.logout()
                 break
             else:
                 print("Invalid option. Try again.")
 
-    def __welcome_user(self):
+    def _welcome_user(self):
         user = self.get_user()
         if user and 'username' in user:
             now = datetime.now()
             print(f"\nWelcome {user['username']}! Date: {now.strftime('%d-%b-%Y')} Time: {now.strftime('%I:%M%p')}")
         else:
-            print("⚠️ User not logged in.")
-
-    def __headlines_menu(self):
-        response = True
-        while True and response:
-            choice = print_menu("Headlines", ["Today", "Date range", "Back"])
-            if choice == '1':
-                response = self.__show_headlines(str(date.today()), str(date.today()))
-            elif choice == '2':
-                start, end = get_date()
-                response = self.__show_headlines(start, end)
-            elif choice == '3':
-                break
-            else:
-                print("Invalid option.")
-        
-        return response
-
-
-    def __show_headlines(self, start_date, end_date):
-        categories_resp = self.api_client.make_request('GET', 'user/categories', current_user=self.get_user())
-        categories = categories_resp.get("categories", [])
-        category_map = {'1': None}
-        print("\n--- Select a Category ---")
-        print("1. All")
-        for idx, cat in enumerate(categories, start=2):
-            print(f"{idx}. {cat['name']}")
-            category_map[str(idx)] = cat['name']
-        print(f"{len(categories) + 2}. Back")
-
-        choice = input("Select an option: ").strip()
-        if choice == str(len(categories) + 2):
-            return True
-
-        selected = category_map.get(choice)
-        
-        params = {'start_date': start_date, 'end_date': end_date}
-        if selected:
-            params['category'] = selected
-
-        response = self.api_client.make_request('GET', 'user/news', params=params, current_user=self.get_user())
-        articles = response.get('articles', []) if response.get('success') else []
-        display_articles(articles)
-
-        if articles:
-            response = self.__article_interaction_loop()
-            return response
-        
-        return True
-
-    def __article_interaction_loop(self):
-        response = True
-        while True:
-            choice = print_menu("Article Options", ["Back", "Logout", "Save Article", "Like/Dislike Article", "Report Article"])
-            if choice == '1':
-                break
-            elif choice == '2':
-                self.user_service.logout()
-                return False
-            elif choice in ['3', '4']:
-                self.__handle_like_dislike(choice)
-                
-            elif choice in ["5"]:
-                self.__handle_report_article()
-            else:
-                print("Invalid option.")
-
-        return response
-    
-    def __handle_like_dislike(self, choice):
-        article_id = get_valid_article_id()
-        if article_id is not None:
-            if choice == '3':
-                resp = self.article_service.save_article(article_id)
-                print(resp.get('message', "Failed to save article."))
-            else:
-                self.__react_to_article(article_id)
-    
-    def __handle_report_article(self):
-        article_id = get_valid_article_id()
-        report_reason  = input("Enter the article report reason\n")
-        if article_id is not None:
-                resp = self.article_service.report_article(article_id, report_reason)
-                print(resp)
-        else:
-            print("Artical Reporting Failed")
-
-
-    
-    def __react_to_article(self, article_id):
-        reaction = input("1. Like\n2. Dislike\nChoose reaction: ").strip()
-        if reaction == '1':
-            result = self.article_service.react_to_article(article_id, "like")
-        elif reaction == '2':
-            result = self.article_service.react_to_article(article_id, "dislike")
-        else:
-            print("Invalid reaction.")
-            return
-        print(result.get('message', "Failed to react to article."))
-
-
-    def __saved_articles_menu(self):
-        print("\n--- Saved Articles ---")
-        response = self.article_service.get_saved_articles()
-        if isinstance(response, list) and response[0].get('success'):
-            display_articles(response[0].get('articles', []))
-        else:
-            print("Failed to load saved articles.")
-            return
-
-        while True:
-            choice = print_menu("Manage Saved Articles", ["Back", "Logout", "Delete Article"])
-            if choice == '1':
-                break
-            elif choice == '2':
-                self.user_service.logout()
-                return False
-            elif choice == '3':
-                article_id = get_valid_article_id()
-                if article_id is not None:
-                    result = self.article_service.delete_article(article_id)
-                    print(result.get('message', "Failed to delete article."))
-            else:
-                print("Invalid option.")
-
-    def __search_menu(self):
-        print("\n--- Search News ---")
-        query = input("Enter keyword(s): ").strip()
-        if not query:
-            print("Search query can't be empty.")
-            return
-
-        start, end = get_date()
-        params = {
-            'q': query, 'limit': 20, 'offset': 0,
-            'start_date': start.strftime('%Y-%m-%d'),
-            'end_date': end.strftime('%Y-%m-%d')
-        }
-        response = self.api_client.make_request('GET', 'user/news', params=params, current_user=self.get_user())
-        if response.get('success'):
-            articles = response.get('articles', [])
-            display_articles(articles)
-            if articles:
-                self.__article_interaction_loop()
-            else:
-                print("No results found.")
-        else:
-            print("Search failed.")
-
-    def __notifications_menu(self):
-        response = True
-        while True and response:
-            choice = print_menu("Notifications", ["View", "Configure", "Back", "Logout"])
-            if choice == '1':
-                self.__view_notifications()
-            elif choice == '2':
-                response = self.__configure_notifications()
-            elif choice == '3':
-                return True
-            elif choice == '4':
-                self.user_service.logout()
-                return False
-            else:
-                print("Invalid option.")
-                
-        return True
-
-    def __view_notifications(self):
-        response = self.notification_service.get_notifications()
-        if response.get('success'):
-            notifications = response.get('notifications', [])
-            if not notifications:
-                print("No notifications found.")
-                return
-            for i, n in enumerate(notifications, 1):
-                print(f"{i}. Category ID: {n.get('category_id')}, Articles: {n.get('article_ids')}")
-                print(f"   Message: {n.get('message')}")
-                print(f"   Sent At: {n.get('sent_at')}")
-                print("-" * 40)
-        else:
-            print("Failed to load notifications.")
-
-    def __configure_notifications(self):
-        while True:
-            prefs = self.notification_service.get_user_preferences()
-            categories_resp = self.api_client.make_request('GET', 'user/categories', current_user=self.get_user())
-            if not (prefs and prefs.get('success') and categories_resp.get('success')):
-                print("Failed to load preferences or categories.")
-                return
-
-            categories = categories_resp.get("categories", [])
-            enabled = {p['name']: p['enabled'] for p in prefs['preferences'].get('category_preferences', [])}
-
-            print("\n--- Notification Categories ---")
-            for idx, cat in enumerate(categories, 1):
-                status = "Enabled" if enabled.get(cat['name'], False) else "Disabled"
-                print(f"{idx}. {cat['name']} - {status}")
-            print(f"{len(categories) + 1}. Manage Keywords")
-            print(f"{len(categories) + 2}. Back")
-            print(f"{len(categories) + 3}. Logout")
-
-            option = input("Choose an option: ").strip()
-            try:
-                opt = int(option)
-                if 1 <= opt <= len(categories):
-                    cat_name = categories[opt - 1]['name']
-                    resp = self.notification_service.toggle_category_notification(cat_name)
-                    print(resp.get('message', "Failed to update setting."))
-                elif opt == len(categories) + 1:
-                    self.__manage_keywords()
-                elif opt == len(categories) + 2:
-                    return True
-                elif opt == len(categories) + 3:
-                    self.user_service.logout()
-                    return False
-                else:
-                    print("Invalid option.")
-            except ValueError:
-                print("Invalid input.")
-        
-        return True
-
-    def __manage_keywords(self):
-        prefs = self.notification_service.get_user_preferences()
-        category_prefs = self._extract_category_preferences(prefs)
-        
-        if not category_prefs:
-            print("no Categories Found")
-            return
-
-        selected_category = self._select_category(category_prefs)
-        if not selected_category:
-            return
-
-        self._handle_category_action(selected_category)
-    
-    
-    def _extract_category_preferences(self, prefs):
-        return prefs['preferences'].get('category_preferences', [])
-
-    
-    def _select_category(self, categories):
-        self._print_category_list(categories)
-        try:
-            idx = int(input("Select a category number: ")) - 1
-            return categories[idx]
-        except (ValueError, IndexError):
-            print("Invalid category selection.")
-            return None
-    
-    
-    
-    def _print_category_list(self, categories):
-        print("\nAvailable Categories:")
-        for idx, cat in enumerate(categories, 1):
-            print(f"{idx}. {cat['name']}")
-    
-    
-    def _handle_category_action(self, category):
-        self._print_category_details(category)
-        choice = print_menu("Manage Keywords", ["Add Keyword(s)", "Remove Keyword(s)", "Disable Category", "Back"])
-
-        if choice == '1':
-            self._handle_add_keywords(category)
-        elif choice == '2':
-            self._handle_remove_keywords(category)
-        elif choice == '3':
-            self._handle_disable_category(category)
-    
-    
-    def _print_category_details(self, category):
-        keywords = category.get('keywords', [])
-        enabled = category.get('enabled', True)
-        print(f"\nSelected Category: {category['name']}")
-        print(f"Enabled: {'Yes' if enabled else 'No'}")
-        print(f"Current Keywords: {', '.join(keywords) if keywords else 'None'}")
-    
-    
-    def _handle_add_keywords(self, category):
-        new_keywords = self._get_keywords_input("Enter comma-separated keywords to add: ")
-        if not new_keywords:
-            print("No keywords entered.")
-            return
-        resp = self.notification_service.update_keywords(category['name'], new_keywords)
-        print(resp.get('message', 'Failed to add keywords.'))
-    
-    
-    
-    def _handle_remove_keywords(self, category):
-        keywords = category.get('keywords', [])
-        if not keywords:
-            print("No keywords to remove.")
-            return
-
-        self._print_keyword_list(keywords)
-        selected_keywords = self._get_selected_keywords(keywords)
-        if not selected_keywords:
-            print("No valid keywords selected.")
-            return
-
-        
-        resp = self.notification_service.delete_keyword(category['name'], selected_keywords)
-        print(resp.get('message', 'Failed to remove keywords.'))
-        
-    
-    
-    def _print_keyword_list(self, keywords):
-        print("\nKeywords:")
-        for idx, kw in enumerate(keywords, 1):
-            print(f"{idx}. {kw}")
-
-    
-    def _get_selected_keywords(self, keywords):
-        try:
-            indices = input("Enter keyword numbers to remove (comma-separated): ")
-            idx_list = [int(i.strip()) - 1 for i in indices.split(',')]
-            return [keywords[i] for i in idx_list if 0 <= i < len(keywords)]
-        except (ValueError, IndexError):
-            return []
-    
-    
-    def _get_keywords_input(self, prompt_msg):
-        raw = input(prompt_msg).strip().lower().split(',')
-        return [k.strip() for k in raw if k.strip()]
-
-    
-    def _handle_disable_category(self, category):
-        confirm = input(f"Disable '{category['name']}' and clear keywords? (y/n): ").strip().lower()
-        if confirm == 'y':
-            resp = self.notification_service.delete_keyword(category['name'], [])
-            print(resp.get('message', 'Failed to disable category.'))
-        else:
-            print("Cancelled.")
-
-
-
-
-
-
-
-
-
+            print("User not logged in.")
