+++ source_code/server/routes/user_like_routes.py
@@ -1,12 +1,14 @@
 from flask import Blueprint, request, jsonify
 from services.article_reaction import ArticleReactionService
-from models.article_reaction import ArticleReactionModel
-from database.database import db
+from repository.article_reaction import ArticleReactionRepository
+
+from database.database import db 
+
 
 article_reaction_bp = Blueprint('article_reaction', __name__)
 
 def get_service():
-    repo = ArticleReactionModel(db)
+    repo = ArticleReactionRepository(db)
     return ArticleReactionService(repo)
 
 
@@ -20,10 +22,11 @@ def react_to_article(article_id):
     try:
         service.react(user_id, article_id, reaction)
         return jsonify({"success": True,"message": f"{reaction} recorded successfully"}), 200
-    except ValueError as ve:
-        return jsonify({"success": False,"error": str(ve)}), 400
-    except Exception as e:
+    except ValueError as error:
+        return jsonify({"success": False,"error": str(error)}), 400
+    except Exception:
         return jsonify({"success": False,"error": "Something went wrong"}), 500
+
 
 @article_reaction_bp.route('/articles/<int:article_id>/<string:reaction_type>', methods=['GET'])
 def get_reaction_count(article_id, reaction_type):
@@ -31,5 +34,5 @@ def get_reaction_count(article_id, reaction_type):
     try:
         count = service.get_count(article_id, reaction_type)
         return jsonify({"success": True,"article_id": article_id, reaction_type: count}), 200
-    except Exception as error:
+    except Exception:
         return jsonify({"success": False,"error": "Something went wrong"}), 500
